// apps/jobs/web/src/lib/api/generated/model/job.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire Jobs API
 * Jobs mini-app API
 * OpenAPI spec version: 0.1.0
 */
import type { CompanyRef } from './companyRef';
import type { JobFunction } from './jobFunction';
import type { JobSeniority } from './jobSeniority';
import type { JobEmploymentType } from './jobEmploymentType';
import type { JobWorkplace } from './jobWorkplace';
import type { JobLocation } from './jobLocation';
import type { JobSalary } from './jobSalary';
import type { JobApplyMode } from './jobApplyMode';
import type { JobScreeningQuestionsItem } from './jobScreeningQuestionsItem';
import type { JobStatus } from './jobStatus';
import type { JobViewer } from './jobViewer';

export interface Job {
  id: string;
  slug: string;
  title: string;
  description: string;
  company: CompanyRef;
  function: JobFunction;
  seniority: JobSeniority;
  employmentType: JobEmploymentType;
  workplace: JobWorkplace;
  /** @nullable */
  location: JobLocation;
  /** @nullable */
  salary: JobSalary;
  skills: string[];
  applyMode: JobApplyMode;
  /** @nullable */
  externalApplyUrl: string | null;
  screeningQuestions: JobScreeningQuestionsItem[];
  status: JobStatus;
  /** @nullable */
  rejectionReason: string | null;
  /** @nullable */
  publishedAt: string | null;
  /** @nullable */
  expiresAt: string | null;
  applicationCount: number;
  createdAt: string;
  viewer: JobViewer;
}